+: ansible scripts to fetch and deploy application environment files
I still need to update the deploy.sh.
But since currently all servers fetch the master branch, that would
break the deploy of existing AWS environments that are not yet using
configuration management.

I will make the needed changes after the new ansible cookbook is deployed
in which we made the change to fetch tag versions instead of the master branch