+: deploy admin/webnode now compatible with new flags for composer and no-upgrade
The old versions had all kinds of checks to see how parameters were provided on
the CLI. This was not maintenable, especially when adding 2 additional flags

I've removed all the where clauses regarding flag checking for specific values.
All flags provided on the CLI are now directly passed through, with a max
of 4 parameters.